Removed comments
